FOR_UPSTREAM [VPG]: drm/i915: Support for -ve delta value for relocation target buffer
The delta value, which has to be added to the base address of the
target buffer of relocation, is currently defined as u32. So only
positive delta values could be passed. To mitigate a Hw issue on CHV,
a Sw WA is required to be implemented, in which 3D Driver requires
the support to pass the negative values also for the delta.
To fulfill that, KMD will internally treat the passed delta value as a signed
value, typecast into s32 and do the arithemtic.
So a negative value will get passed as a large postive u32 number, but
this will also reduce the max value of delta from (2^32 - 1) to
(2^31 - 1), which is not an issue as per the 3D team.

Issue: GMINL-7441

+++ drivers/gpu/drm/i915/i915_gem_execbuffer.c
@@ -259,7 +259,7 @@ relocate_entry_cpu(struct drm_i915_gem_object *obj,
 {
 	struct drm_device *dev = obj->base.dev;
 	uint32_t page_offset = offset_in_page(reloc->offset);
-	uint64_t delta = reloc->delta + target_offset;
+	uint64_t delta = (int32_t)reloc->delta + target_offset;
 	char *vaddr;
 	int ret;
 
@@ -295,7 +295,7 @@ relocate_entry_gtt(struct drm_i915_gem_object *obj,
 {
 	struct drm_device *dev = obj->base.dev;
 	struct drm_i915_private *dev_priv = dev->dev_private;
-	uint64_t delta = reloc->delta + target_offset;
+	uint64_t delta = (int32_t)reloc->delta + target_offset;
 	uint32_t __iomem *reloc_entry;
 	void __iomem *reloc_page;
 	int ret;
